+++ gateway/src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
 
     @DeleteMapping("/{userId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable @NonNull Integer userId) {
+    public void deleteUser(@PathVariable @NonNull Long userId) {
         log.info("Gateway: Received request to delete user with id= {}", userId);
         userClient.deleteUser(userId);
     }
