+++ server/src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
 
     @DeleteMapping("/{userId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Integer userId) {
+    public void deleteUser(@PathVariable Long userId) {
         log.info("Server: Received request to delete user with id= {}", userId);
         userService.deleteUser(userId);
     }
